rename and move wailing arrow to talent
